+++ ashinamo/appdata/views.py
@@ -6,29 +6,28 @@
 import simplejson as json
 
 from django.http import HttpResponse
-from ashinamo.lib.class_data import \
-    CpuClassData, NetClassData, IoClassData, MemClassData
+from ashinamo.lib.class_based import cpu, net, io, mem
 
 
 def getcpu(request):
-    cpudata = CpuClassData.CpuData()
+    cpudata = cpu.CpuData()
     result = json.dumps(cpudata.compute_data())
     return HttpResponse(result)
 
 
 def getmem(request):
-    memdata = MemClassData.MemData()
+    memdata = mem.MemData()
     result = json.dumps(memdata.get_data())
     return HttpResponse(result)
 
 
 def getnet(request):
-    netdata = NetClassData.NetData(['eth0'])
+    netdata = net.NetData(['eth0'])
     result = json.dumps(netdata.compute_data())
     return HttpResponse(result)
 
 
 def getio(request):
-    iodata = IoClassData.IoData(['sda'])
+    iodata = io.IoData(['sda'])
     result = json.dumps(iodata.compute_data())
     return HttpResponse(result)
