+++ ashinamo/lib/function_based/cpu.py
