+++ ashinamo/lib/function_based/mem.py
