+++ ashinamo/lib/shell_based/cpu.py
