+++ ashinamo/lib/shell_based/io.py
